Tree: Resize map memory to the exact VXL memory size
Before we would output memory to 10MiB regardless
of the actual memory size.

Most of the time we would fit in this size but its
easy to hit this limit with much more complex maps.
Now we allocate the maximum possible memory size for
given XYZ size of VXL and after filling this memory
with the map info reallocate to the exact map size
of the output VXL memory. And compress that.

This has multiple advantages.
Its impossible to overflow if given valid VXL map
info about the size.
We send the minimal amount of data to player so
faster map loading and compressing.
We can load any size of valid VXL map. Even those
that may be currently non existent. Such as
1024x1024x256.

+++ Source/Structs.h
@@ -80,12 +80,12 @@ typedef struct
     uint8       mapCount;
     stringNode* currentMap;
     // compressed map
-    Queue*            compressedMap;
-    uint32            compressedSize;
-    Vector3i          resultLine[50];
-    long unsigned int mapSize;
-    MapVxl            map;
-    stringNode*       mapList;
+    Queue*      compressedMap;
+    uint32      compressedSize;
+    Vector3i    resultLine[50];
+    size_t      mapSize;
+    MapVxl      map;
+    stringNode* mapList;
 } Map;
 
 typedef struct
